Refactor 'globals.d.ts'
- Move 'timers'-specific types to `timers.d.ts`
- Move 'buffer'-specific types to `buffer.d.ts`
- Move 'errors'-specific types to `errors.d.ts` (new file)
- Move 'stream'-specific types to `stream.d.ts`
- Add `URL` and `URLSearchParams` globals in `url.d.ts`
- Add `TextDecoder` and `TextEncoder` globals in `util.d.ts`
- Make `gs` optional
- Add himself to `index.d.ts` for receiving notifications about new `node` issues
