+++ types/node/globals.d.ts
@@ -1,18 +1,3 @@
-// Declare "static" methods in Error
-interface ErrorConstructor {
-    /** Create .stack property on a target object */
-    captureStackTrace(targetObject: object, constructorOpt?: Function): void;
-
-    /**
-     * Optional override for formatting stack traces
-     *
-     * @see https://github.com/v8/v8/wiki/Stack%20Trace%20API#customizing-stack-traces
-     */
-    prepareStackTrace?: (err: Error, stackTraces: NodeJS.CallSite[]) => any;
-
-    stackTraceLimit: number;
-}
-
 // Node.js ESNEXT support
 interface String {
     /** Removes whitespace from the left end of a string. */
@@ -47,21 +32,6 @@ declare var console: Console;
 declare var __filename: string;
 declare var __dirname: string;
 
-declare function setTimeout(callback: (...args: any[]) => void, ms: number, ...args: any[]): NodeJS.Timeout;
-declare namespace setTimeout {
-    function __promisify__(ms: number): Promise<void>;
-    function __promisify__<T>(ms: number, value: T): Promise<T>;
-}
-declare function clearTimeout(timeoutId: NodeJS.Timeout): void;
-declare function setInterval(callback: (...args: any[]) => void, ms: number, ...args: any[]): NodeJS.Timeout;
-declare function clearInterval(intervalId: NodeJS.Timeout): void;
-declare function setImmediate(callback: (...args: any[]) => void, ...args: any[]): NodeJS.Immediate;
-declare namespace setImmediate {
-    function __promisify__(): Promise<void>;
-    function __promisify__<T>(value: T): Promise<T>;
-}
-declare function clearImmediate(immediateId: NodeJS.Immediate): void;
-
 declare function queueMicrotask(callback: () => void): void;
 
 declare var require: NodeRequire;
@@ -69,247 +39,6 @@ declare var module: NodeModule;
 
 // Same as module.exports
 declare var exports: any;
-
-// Buffer class
-type BufferEncoding = "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary" | "hex";
-
-type WithImplicitCoercion<T> = T | { valueOf(): T };
-
-/**
- * Raw data is stored in instances of the Buffer class.
- * A Buffer is similar to an array of integers but corresponds to a raw memory allocation outside the V8 heap.  A Buffer cannot be resized.
- * Valid string encodings: 'ascii'|'utf8'|'utf16le'|'ucs2'(alias of 'utf16le')|'base64'|'binary'(deprecated)|'hex'
- */
-declare class Buffer extends Uint8Array {
-    /**
-     * Allocates a new buffer containing the given {str}.
-     *
-     * @param str String to store in buffer.
-     * @param encoding encoding to use, optional.  Default is 'utf8'
-     * @deprecated since v10.0.0 - Use `Buffer.from(string[, encoding])` instead.
-     */
-    constructor(str: string, encoding?: BufferEncoding);
-    /**
-     * Allocates a new buffer of {size} octets.
-     *
-     * @param size count of octets to allocate.
-     * @deprecated since v10.0.0 - Use `Buffer.alloc()` instead (also see `Buffer.allocUnsafe()`).
-     */
-    constructor(size: number);
-    /**
-     * Allocates a new buffer containing the given {array} of octets.
-     *
-     * @param array The octets to store.
-     * @deprecated since v10.0.0 - Use `Buffer.from(array)` instead.
-     */
-    constructor(array: Uint8Array);
-    /**
-     * Produces a Buffer backed by the same allocated memory as
-     * the given {ArrayBuffer}/{SharedArrayBuffer}.
-     *
-     *
-     * @param arrayBuffer The ArrayBuffer with which to share memory.
-     * @deprecated since v10.0.0 - Use `Buffer.from(arrayBuffer[, byteOffset[, length]])` instead.
-     */
-    constructor(arrayBuffer: ArrayBuffer | SharedArrayBuffer);
-    /**
-     * Allocates a new buffer containing the given {array} of octets.
-     *
-     * @param array The octets to store.
-     * @deprecated since v10.0.0 - Use `Buffer.from(array)` instead.
-     */
-    constructor(array: ReadonlyArray<any>);
-    /**
-     * Copies the passed {buffer} data onto a new {Buffer} instance.
-     *
-     * @param buffer The buffer to copy.
-     * @deprecated since v10.0.0 - Use `Buffer.from(buffer)` instead.
-     */
-    constructor(buffer: Buffer);
-    /**
-     * When passed a reference to the .buffer property of a TypedArray instance,
-     * the newly created Buffer will share the same allocated memory as the TypedArray.
-     * The optional {byteOffset} and {length} arguments specify a memory range
-     * within the {arrayBuffer} that will be shared by the Buffer.
-     *
-     * @param arrayBuffer The .buffer property of any TypedArray or a new ArrayBuffer()
-     */
-    static from(arrayBuffer: WithImplicitCoercion<ArrayBuffer | SharedArrayBuffer>, byteOffset?: number, length?: number): Buffer;
-    /**
-     * Creates a new Buffer using the passed {data}
-     * @param data data to create a new Buffer
-     */
-    static from(data: Uint8Array | ReadonlyArray<number>): Buffer;
-    static from(data: WithImplicitCoercion<Uint8Array | ReadonlyArray<number> | string>): Buffer;
-    /**
-     * Creates a new Buffer containing the given JavaScript string {str}.
-     * If provided, the {encoding} parameter identifies the character encoding.
-     * If not provided, {encoding} defaults to 'utf8'.
-     */
-    static from(str: WithImplicitCoercion<string> | { [Symbol.toPrimitive](hint: 'string'): string }, encoding?: BufferEncoding): Buffer;
-    /**
-     * Creates a new Buffer using the passed {data}
-     * @param values to create a new Buffer
-     */
-    static of(...items: number[]): Buffer;
-    /**
-     * Returns true if {obj} is a Buffer
-     *
-     * @param obj object to test.
-     */
-    static isBuffer(obj: any): obj is Buffer;
-    /**
-     * Returns true if {encoding} is a valid encoding argument.
-     * Valid string encodings in Node 0.12: 'ascii'|'utf8'|'utf16le'|'ucs2'(alias of 'utf16le')|'base64'|'binary'(deprecated)|'hex'
-     *
-     * @param encoding string to test.
-     */
-    static isEncoding(encoding: string): encoding is BufferEncoding;
-    /**
-     * Gives the actual byte length of a string. encoding defaults to 'utf8'.
-     * This is not the same as String.prototype.length since that returns the number of characters in a string.
-     *
-     * @param string string to test.
-     * @param encoding encoding used to evaluate (defaults to 'utf8')
-     */
-    static byteLength(
-        string: string | NodeJS.ArrayBufferView | ArrayBuffer | SharedArrayBuffer,
-        encoding?: BufferEncoding
-    ): number;
-    /**
-     * Returns a buffer which is the result of concatenating all the buffers in the list together.
-     *
-     * If the list has no items, or if the totalLength is 0, then it returns a zero-length buffer.
-     * If the list has exactly one item, then the first item of the list is returned.
-     * If the list has more than one item, then a new Buffer is created.
-     *
-     * @param list An array of Buffer objects to concatenate
-     * @param totalLength Total length of the buffers when concatenated.
-     *   If totalLength is not provided, it is read from the buffers in the list. However, this adds an additional loop to the function, so it is faster to provide the length explicitly.
-     */
-    static concat(list: ReadonlyArray<Uint8Array>, totalLength?: number): Buffer;
-    /**
-     * The same as buf1.compare(buf2).
-     */
-    static compare(buf1: Uint8Array, buf2: Uint8Array): number;
-    /**
-     * Allocates a new buffer of {size} octets.
-     *
-     * @param size count of octets to allocate.
-     * @param fill if specified, buffer will be initialized by calling buf.fill(fill).
-     *    If parameter is omitted, buffer will be filled with zeros.
-     * @param encoding encoding used for call to buf.fill while initalizing
-     */
-    static alloc(size: number, fill?: string | Buffer | number, encoding?: BufferEncoding): Buffer;
-    /**
-     * Allocates a new buffer of {size} octets, leaving memory not initialized, so the contents
-     * of the newly created Buffer are unknown and may contain sensitive data.
-     *
-     * @param size count of octets to allocate
-     */
-    static allocUnsafe(size: number): Buffer;
-    /**
-     * Allocates a new non-pooled buffer of {size} octets, leaving memory not initialized, so the contents
-     * of the newly created Buffer are unknown and may contain sensitive data.
-     *
-     * @param size count of octets to allocate
-     */
-    static allocUnsafeSlow(size: number): Buffer;
-    /**
-     * This is the number of bytes used to determine the size of pre-allocated, internal Buffer instances used for pooling. This value may be modified.
-     */
-    static poolSize: number;
-
-    write(string: string, encoding?: BufferEncoding): number;
-    write(string: string, offset: number, encoding?: BufferEncoding): number;
-    write(string: string, offset: number, length: number, encoding?: BufferEncoding): number;
-    toString(encoding?: BufferEncoding, start?: number, end?: number): string;
-    toJSON(): { type: 'Buffer'; data: number[] };
-    equals(otherBuffer: Uint8Array): boolean;
-    compare(
-        otherBuffer: Uint8Array,
-        targetStart?: number,
-        targetEnd?: number,
-        sourceStart?: number,
-        sourceEnd?: number
-    ): number;
-    copy(targetBuffer: Uint8Array, targetStart?: number, sourceStart?: number, sourceEnd?: number): number;
-    /**
-     * Returns a new `Buffer` that references **the same memory as the original**, but offset and cropped by the start and end indices.
-     *
-     * This method is incompatible with `Uint8Array#slice()`, which returns a copy of the original memory.
-     *
-     * @param begin Where the new `Buffer` will start. Default: `0`.
-     * @param end Where the new `Buffer` will end (not inclusive). Default: `buf.length`.
-     */
-    slice(begin?: number, end?: number): Buffer;
-    /**
-     * Returns a new `Buffer` that references **the same memory as the original**, but offset and cropped by the start and end indices.
-     *
-     * This method is compatible with `Uint8Array#subarray()`.
-     *
-     * @param begin Where the new `Buffer` will start. Default: `0`.
-     * @param end Where the new `Buffer` will end (not inclusive). Default: `buf.length`.
-     */
-    subarray(begin?: number, end?: number): Buffer;
-    writeBigInt64BE(value: bigint, offset?: number): number;
-    writeBigInt64LE(value: bigint, offset?: number): number;
-    writeBigUInt64BE(value: bigint, offset?: number): number;
-    writeBigUInt64LE(value: bigint, offset?: number): number;
-    writeUIntLE(value: number, offset: number, byteLength: number): number;
-    writeUIntBE(value: number, offset: number, byteLength: number): number;
-    writeIntLE(value: number, offset: number, byteLength: number): number;
-    writeIntBE(value: number, offset: number, byteLength: number): number;
-    readBigUInt64BE(offset?: number): bigint;
-    readBigUInt64LE(offset?: number): bigint;
-    readBigInt64BE(offset?: number): bigint;
-    readBigInt64LE(offset?: number): bigint;
-    readUIntLE(offset: number, byteLength: number): number;
-    readUIntBE(offset: number, byteLength: number): number;
-    readIntLE(offset: number, byteLength: number): number;
-    readIntBE(offset: number, byteLength: number): number;
-    readUInt8(offset?: number): number;
-    readUInt16LE(offset?: number): number;
-    readUInt16BE(offset?: number): number;
-    readUInt32LE(offset?: number): number;
-    readUInt32BE(offset?: number): number;
-    readInt8(offset?: number): number;
-    readInt16LE(offset?: number): number;
-    readInt16BE(offset?: number): number;
-    readInt32LE(offset?: number): number;
-    readInt32BE(offset?: number): number;
-    readFloatLE(offset?: number): number;
-    readFloatBE(offset?: number): number;
-    readDoubleLE(offset?: number): number;
-    readDoubleBE(offset?: number): number;
-    reverse(): this;
-    swap16(): Buffer;
-    swap32(): Buffer;
-    swap64(): Buffer;
-    writeUInt8(value: number, offset?: number): number;
-    writeUInt16LE(value: number, offset?: number): number;
-    writeUInt16BE(value: number, offset?: number): number;
-    writeUInt32LE(value: number, offset?: number): number;
-    writeUInt32BE(value: number, offset?: number): number;
-    writeInt8(value: number, offset?: number): number;
-    writeInt16LE(value: number, offset?: number): number;
-    writeInt16BE(value: number, offset?: number): number;
-    writeInt32LE(value: number, offset?: number): number;
-    writeInt32BE(value: number, offset?: number): number;
-    writeFloatLE(value: number, offset?: number): number;
-    writeFloatBE(value: number, offset?: number): number;
-    writeDoubleLE(value: number, offset?: number): number;
-    writeDoubleBE(value: number, offset?: number): number;
-
-    fill(value: string | Uint8Array | number, offset?: number, end?: number, encoding?: BufferEncoding): this;
-
-    indexOf(value: string | number | Uint8Array, byteOffset?: number, encoding?: BufferEncoding): number;
-    lastIndexOf(value: string | number | Uint8Array, byteOffset?: number, encoding?: BufferEncoding): number;
-    entries(): IterableIterator<[number, number]>;
-    includes(value: string | number | Buffer, byteOffset?: number, encoding?: BufferEncoding): boolean;
-    keys(): IterableIterator<number>;
-    values(): IterableIterator<number>;
-}
 
 /*----------------------------------------------*
 *                                               *
@@ -358,122 +87,12 @@ declare namespace NodeJS {
         sorted?: boolean | ((a: string, b: string) => number);
     }
 
-    interface CallSite {
-        /**
-         * Value of "this"
-         */
-        getThis(): any;
-
-        /**
-         * Type of "this" as a string.
-         * This is the name of the function stored in the constructor field of
-         * "this", if available.  Otherwise the object's [[Class]] internal
-         * property.
-         */
-        getTypeName(): string | null;
-
-        /**
-         * Current function
-         */
-        getFunction(): Function | undefined;
-
-        /**
-         * Name of the current function, typically its name property.
-         * If a name property is not available an attempt will be made to try
-         * to infer a name from the function's context.
-         */
-        getFunctionName(): string | null;
-
-        /**
-         * Name of the property [of "this" or one of its prototypes] that holds
-         * the current function
-         */
-        getMethodName(): string | null;
-
-        /**
-         * Name of the script [if this function was defined in a script]
-         */
-        getFileName(): string | null;
-
-        /**
-         * Current line number [if this function was defined in a script]
-         */
-        getLineNumber(): number | null;
-
-        /**
-         * Current column number [if this function was defined in a script]
-         */
-        getColumnNumber(): number | null;
-
-        /**
-         * A call site object representing the location where eval was called
-         * [if this function was created using a call to eval]
-         */
-        getEvalOrigin(): string | undefined;
-
-        /**
-         * Is this a toplevel invocation, that is, is "this" the global object?
-         */
-        isToplevel(): boolean;
-
-        /**
-         * Does this call take place in code defined by a call to eval?
-         */
-        isEval(): boolean;
-
-        /**
-         * Is this call in native V8 code?
-         */
-        isNative(): boolean;
-
-        /**
-         * Is this a constructor call?
-         */
-        isConstructor(): boolean;
-    }
-
-    interface ErrnoException extends Error {
-        errno?: number;
-        code?: string;
-        path?: string;
-        syscall?: string;
-        stack?: string;
-    }
-
-    interface ReadableStream extends EventEmitter {
-        readable: boolean;
-        read(size?: number): string | Buffer;
-        setEncoding(encoding: BufferEncoding): this;
-        pause(): this;
-        resume(): this;
-        isPaused(): boolean;
-        pipe<T extends WritableStream>(destination: T, options?: { end?: boolean; }): T;
-        unpipe(destination?: WritableStream): this;
-        unshift(chunk: string | Uint8Array, encoding?: BufferEncoding): void;
-        wrap(oldStream: ReadableStream): this;
-        [Symbol.asyncIterator](): AsyncIterableIterator<string | Buffer>;
-    }
-
-    interface WritableStream extends EventEmitter {
-        writable: boolean;
-        write(buffer: Uint8Array | string, cb?: (err?: Error | null) => void): boolean;
-        write(str: string, encoding?: BufferEncoding, cb?: (err?: Error | null) => void): boolean;
-        end(cb?: () => void): void;
-        end(data: string | Uint8Array, cb?: () => void): void;
-        end(str: string, encoding?: BufferEncoding, cb?: () => void): void;
-    }
-
-    interface ReadWriteStream extends ReadableStream, WritableStream { }
-
     interface Global {
         Array: typeof Array;
         ArrayBuffer: typeof ArrayBuffer;
         Boolean: typeof Boolean;
-        Buffer: typeof Buffer;
         DataView: typeof DataView;
         Date: typeof Date;
-        Error: typeof Error;
-        EvalError: typeof EvalError;
         Float32Array: typeof Float32Array;
         Float64Array: typeof Float64Array;
         Function: typeof Function;
@@ -489,24 +108,16 @@ declare namespace NodeJS {
         Number: typeof Number;
         Object: typeof Object;
         Promise: typeof Promise;
-        RangeError: typeof RangeError;
-        ReferenceError: typeof ReferenceError;
         RegExp: typeof RegExp;
         Set: SetConstructor;
         String: typeof String;
         Symbol: Function;
-        SyntaxError: typeof SyntaxError;
-        TypeError: typeof TypeError;
-        URIError: typeof URIError;
         Uint16Array: typeof Uint16Array;
         Uint32Array: typeof Uint32Array;
         Uint8Array: typeof Uint8Array;
         Uint8ClampedArray: typeof Uint8ClampedArray;
         WeakMap: WeakMapConstructor;
         WeakSet: WeakSetConstructor;
-        clearImmediate: (immediateId: Immediate) => void;
-        clearInterval: (intervalId: Timeout) => void;
-        clearTimeout: (timeoutId: Timeout) => void;
         decodeURI: typeof decodeURI;
         decodeURIComponent: typeof decodeURIComponent;
         encodeURI: typeof encodeURI;
@@ -518,37 +129,11 @@ declare namespace NodeJS {
         isNaN: typeof isNaN;
         parseFloat: typeof parseFloat;
         parseInt: typeof parseInt;
-        setImmediate: (callback: (...args: any[]) => void, ...args: any[]) => Immediate;
-        setInterval: (callback: (...args: any[]) => void, ms: number, ...args: any[]) => Timeout;
-        setTimeout: (callback: (...args: any[]) => void, ms: number, ...args: any[]) => Timeout;
         queueMicrotask: typeof queueMicrotask;
         undefined: typeof undefined;
         unescape: (str: string) => string;
-        gc: () => void;
+        gc?: () => void;
         v8debug?: any;
-    }
-
-    interface RefCounted {
-        ref(): this;
-        unref(): this;
-    }
-
-    // compatibility with older typings
-    interface Timer extends RefCounted {
-        hasRef(): boolean;
-        refresh(): this;
-        [Symbol.toPrimitive](): number;
-    }
-
-    interface Immediate extends RefCounted {
-        hasRef(): boolean;
-        _onImmediate: Function; // to distinguish it from the Timeout class
-    }
-
-    interface Timeout extends Timer {
-        hasRef(): boolean;
-        refresh(): this;
-        [Symbol.toPrimitive](): number;
     }
 
     type TypedArray =
