+++ types/node/stream.d.ts
@@ -351,4 +351,33 @@ declare module "stream" {
     }
 
     export = internal;
+
+    global {
+        namespace NodeJS {
+            interface ReadableStream extends EventEmitter {
+                readable: boolean;
+                read(size?: number): string | Buffer;
+                setEncoding(encoding: BufferEncoding): this;
+                pause(): this;
+                resume(): this;
+                isPaused(): boolean;
+                pipe<T extends WritableStream>(destination: T, options?: { end?: boolean; }): T;
+                unpipe(destination?: WritableStream): this;
+                unshift(chunk: string | Uint8Array, encoding?: BufferEncoding): void;
+                wrap(oldStream: ReadableStream): this;
+                [Symbol.asyncIterator](): AsyncIterableIterator<string | Buffer>;
+            }
+
+            interface WritableStream extends EventEmitter {
+                writable: boolean;
+                write(buffer: Uint8Array | string, cb?: (err?: Error | null) => void): boolean;
+                write(str: string, encoding?: BufferEncoding, cb?: (err?: Error | null) => void): boolean;
+                end(cb?: () => void): void;
+                end(data: string | Uint8Array, cb?: () => void): void;
+                end(str: string, encoding?: BufferEncoding, cb?: () => void): void;
+            }
+
+            interface ReadWriteStream extends ReadableStream, WritableStream { }
+        }
+    }
 }
