+++ types/node/ts3.4/base.d.ts
@@ -15,6 +15,7 @@
 // Base definitions for all NodeJS modules that are not specific to any version of TypeScript:
 
 /// <reference path="../globals.d.ts" />
+/// <reference path="../globals.errors.d.ts" />
 /// <reference path="../async_hooks.d.ts" />
 /// <reference path="../buffer.d.ts" />
 /// <reference path="../child_process.d.ts" />
